chore: 별점순 정렬 수정(했으나 작동 안됨..)

+++ server/src/main/java/com/was_surf/domain/surf_spot/application/SurfSpotService.java
@@ -92,7 +92,7 @@ public class SurfSpotService {
         }
 
         int count = reviews.size();
-        int average = count == 0 ? 0 : sum / count;
+        double average = count == 0 ? 0 : (double)sum / count;
 
         surfSpot.setAverageRating(average);
     }
